+++ {{cookiecutter.project_slug}}/.gitignore
@@ -106,3 +106,112 @@ lightning_logs
 .neptune
 debug_art.log
 *.png
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+*.py,cover
+.hypothesis/
+.pytest_cache/
+cover/
+
+# Jupyter Notebook
+.ipynb_checkpoints/
+
+# IPython
+profile_default/
+ipython_config.py
+
+# Environment directories
+.env
+.venv/
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Pycharm, VSCode, etc.
+.idea/
+.vscode/
+
+# OS-specific
+.DS_Store  # macOS
+Thumbs.db  # Windows
+.dbshell   # Linux (SQLite)
+
+# Other common tools
+*.log
+*.swp
+*.bak
+
+# Django
+*.sqlite3
+local_settings.py
+
+# Flask
+instance/
+.webassets-cache
+
+# Other
+*.pot
+*.pyc
+*.pyo
+*.mo
+*.orig
+*.rej
+*.swp
+*~
+*.bak
+*.tmp
+# TODO: Add more patterns based on your specific needs
+checkpoints
+lightning_logs
+*.pyc
+.neptune
+debug_art.log
+*.png
