--- {{cookiecutter.project_slug}}/models/base_model.py
@@ -1,5 +1,0 @@
-from art.core.base_components import ArtModule
-
-
-class Model(ArtModule):
-    ...
