+++ {{cookiecutter.project_slug}}/models/baselines.py
