+++ {{cookiecutter.project_slug}}/models/mnist_models.py
